requirements.txt
flask==3.0.3
llama-cpp-python==0.2.90
huggingface_hub[cli]==0.23.4
numpy==1.26.4